+++ src/layout/linear/mod.rs
@@ -308,7 +308,7 @@ mod test {
             .add_view(rect)
             .add_view(rect2)
             .arrange()
-            .translate(Point::new(1, 2))
+            .translate_mut(Point::new(1, 2))
             .draw(&mut disp)
             .unwrap();
 
@@ -349,7 +349,7 @@ mod test {
             .add_view(rect)
             .add_view(rect2)
             .arrange()
-            .translate(Point::new(1, 2))
+            .translate_mut(Point::new(1, 2))
             .draw(&mut disp)
             .unwrap();
 
@@ -389,7 +389,7 @@ mod test {
             .add_view(rect)
             .add_view(rect2)
             .arrange()
-            .translate(Point::new(1, 2))
+            .translate_mut(Point::new(1, 2))
             .draw(&mut disp)
             .unwrap();
 
@@ -425,7 +425,7 @@ mod test {
             .add_view(rect)
             .add_view(rect2)
             .arrange()
-            .translate(Point::new(1, 2))
+            .translate_mut(Point::new(1, 2))
             .draw(&mut disp)
             .unwrap();
 
@@ -485,7 +485,7 @@ mod test {
             .add_view(rect)
             .add_view(rect2)
             .arrange()
-            .translate(Point::new(1, 2))
+            .translate_mut(Point::new(1, 2))
             .draw(&mut disp)
             .unwrap();
 
@@ -588,14 +588,14 @@ mod test {
 
     #[test]
     fn layout_size_independent_of_view_location() {
-        let rect = Rectangle::with_size(Point::zero(), Size::new(10, 20));
+        let mut rect = Rectangle::with_size(Point::zero(), Size::new(10, 20));
         let rect2 = Rectangle::with_size(Point::zero(), Size::new(10, 20));
         let size1 = LinearLayout::horizontal()
             .add_view(rect)
             .add_view(rect2)
             .size();
 
-        let rect = rect.translate(Point::new(30, 50));
+        let rect = rect.translate_mut(Point::new(30, 50)).clone();
         let size2 = LinearLayout::horizontal()
             .add_view(rect)
             .add_view(rect2)