+++ src/lib.rs
@@ -185,11 +185,6 @@ pub trait View {
     }
 
     /// Move the origin of an object by a given number of (x, y) pixels,
-    /// by returning a new object
-    #[must_use]
-    fn translate(self, by: Point) -> Self;
-
-    /// Move the origin of an object by a given number of (x, y) pixels,
     /// mutating the object in place
     fn translate_mut(&mut self, by: Point) -> &mut Self;
 
@@ -201,12 +196,6 @@ impl<T> View for T
 where
     T: Transform + Dimensions,
 {
-    #[inline]
-    fn translate(mut self, by: Point) -> Self {
-        Self::translate_mut(&mut self, by);
-        self
-    }
-
     #[inline]
     fn translate_mut(&mut self, by: Point) -> &mut Self {
         Transform::translate_mut(self, by);